requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-adk>=0.1.0
google-genai>=0.3.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

